+++ R/comparisons.R
@@ -92,6 +92,7 @@
 #' @template equivalence
 #' @template type
 #' @template order_of_operations
+#' @template parallel
 #' @template references
 #'
 #' @return A `data.frame` with one row per observation (per term/group) and several columns:
@@ -238,7 +239,7 @@ comparisons <- function(model,
                         conf_level = 0.95,
                         transform = NULL,
                         cross = FALSE,
-                        wts = NULL,
+                        wts = FALSE,
                         hypothesis = NULL,
                         equivalence = NULL,
                         p_adjust = NULL,
@@ -366,7 +367,7 @@ comparisons <- function(model,
         by = by,
         cross = cross,
         comparison = comparison)
-    if (is.null(wts) && "marginaleffects_wts_internal" %in% colnames(newdata)) {
+    if (isFALSE(wts) && "marginaleffects_wts_internal" %in% colnames(newdata)) {
         wts <- "marginaleffects_wts_internal"
     }
 
@@ -591,7 +592,7 @@ avg_comparisons <- function(model,
                             comparison = "difference",
                             transform = NULL,
                             cross = FALSE,
-                            wts = NULL,
+                            wts = FALSE,
                             hypothesis = NULL,
                             equivalence = NULL,
                             p_adjust = NULL,